purchase one stock

+++ db/migrate/20190514210415_create_transactions.rb
@@ -3,7 +3,7 @@ class CreateTransactions < ActiveRecord::Migration[5.2]
     create_table :transactions do |t|
       t.integer :user_id
       t.integer :stock_id
-      t.integer :stock_price
+      t.float :stock_price
       t.integer :quantity
       t.string :date_transaction
